0.2.3 update
Added the mdata test, this test focuses on the testing of the actual .so file rather than the individual modules like previous tests did.

This is essentially a resue of the previous tests, but instead using the .so file directly, rather than compiling the code directly for each test.

Broke up test files to allow for easy reuse of tests.
