+++ src/test/CMakeLists.txt
@@ -7,3 +7,4 @@ add_subdirectory (linky)
 add_subdirectory (stacky)
 add_subdirectory (queue)
 add_subdirectory (binaryTree)
+add_subdirectory (mdata)
