+++ src/test/mdata/mdataTestSuite.h
@@ -1,0 +1,26 @@
+/*
+ * Author -- Meevs
+ * Creation Date -- Wed Dec 29 00:36:57 2021
+ * File Name -- mdataTestSuite.h
+ * Notes -- 
+ */
+
+#ifndef __MDATA_TEST_SUITE__
+#define __MDATA_TEST_SUITE__
+
+#include <stdio.h>
+#include <check.h>
+
+/* Data structure headers */
+#include <linky.h>
+#include <stacky.h>
+#include <queue.h>
+#include <binaryTree.h>
+
+/* Test headers */
+#include <linkyTestSuite.h>
+#include <stackyTestSuite.h>
+#include <queueTestSuite.h>
+#include <binaryTreeTestSuite.h>
+
+#endif
